BKO-113 Update organization of files
